refactor: used computed for location
