+++ package.json
@@ -11,7 +11,7 @@
   },
   "dependencies": {
     "primeicons": "^6.0.1",
-    "primevue": "^3.18.1",
+    "primevue": "^3.21.0",
     "vue": "^3.2.41",
     "vue-router": "^4.1.5"
   },
